==> Rmd/11_pDogLh.Rmd <==
---
title: "**Annex: 11**"
subtitle: "Life History Parameters for Centroscymnus coelolepis"
author: "DELEGATION OF THE EUROPEAN UNION"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: html_document
mathjax: TRUE
fig_width: 6 
fig_height: 4 
tags: SIOFA, deepwater sharks, hierchical framework
license: Creative Commons Attribution-ShareAlike 4.0 International Public License
bibliography: refs.bib
editor_options: 
  markdown: 
    wrap: 72
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir="html")})
---

```{r, knitr, eval=TRUE, echo=FALSE, warning=FALSE, cache=FALSE}
knitr::opts_chunk$set(echo=FALSE)

library(knitr)
opts_chunk$set(cache     =TRUE, 
               comment   =NA, 
               warning   =FALSE, 
               message   =FALSE, 
               error     =FALSE, 
               echo      =FALSE, 
               eval      =TRUE,
               cache     =TRUE,
               cache.path="cache/11/",
               fig.path  ="../figs/11",
               fig.width =6,
               fig.height=6,
               dev       ="png")
iFig=0
iTab=0
```


```{r, libs, cache=FALSE}
library(openxlsx)  # For reading Excel files
library(lubridate) # For handling date and time
library(plyr)      # For data manipulation
library(dplyr)     # For data manipulation
library(reshape)   # For data manipulation
library(ggplot2)   # For plotting
library(ggpubr)    # For plotting
library(mapdata)   # For map-based data
library(viridis)   # For color scales, better for visibility

library(xtable)
library(kableExtra)

library(FLCore)
library(ggplotFL)

library(rfishbase)
library(SPMpriors)
library(FishLife)
```


```{r, theme, cache=FALSE}
theme_my<-function(size=16) {
  theme_minimal() +
    theme(text=element_text(size=size),
          plot.title=element_text(face="bold", size=size+2),
          axis.title=element_text(size=size-1),
          legend.position="bottom",
          legend.title=element_blank(),
          panel.grid.major=element_line(color="gray80"),
          panel.grid.minor=element_blank())}

theme_set(theme_my())  
```


```{r, source}
source("../source/fnBox.R")
source("../source/fnDat.R")
source("../source/fnLbi.R")
source("../source/fl_mcupd_fn.R")
```

```{r}
# Load FAO taxa data
load("../data/inputs/fao_taxa.RData")
```

```{r}
# Read ASFIS species information
asfis = read.csv("../data/inputs/ASFIS_sp_2021.txt")
names(asfis)[4] = "scientificName"
```

```{r}
# Transform fishbase data for compatibility
spp = transform(rfishbase::fishbase, scientificName = paste(Genus, Species))

# Define species of interest for SIOFA
sppSiofa       =c("Centroscymnus coelolepis","Pailona commun","Portugese dogfish","CYO")
sppSiofa       =as.data.frame(t(array(sppSiofa, c(4, length(sppSiofa)/4))))
sppSiofa       =cbind(ldply(strsplit(sppSiofa[,1], " ")), sppSiofa)
names(sppSiofa)=c("Genus", "Species", "scientificName", "french", "name", "FAO")

# Merge species data with fishbase and ASFIS data
sppSiofa = merge(sppSiofa, spp, by=c("Genus", "Species", "scientificName"), all.x=T)
sppSiofa = merge(sppSiofa, asfis[,-(12:13)], by="scientificName", all.x=T)

# Extract life history traits using custom functions
traits = mdply(sppSiofa[,c("Genus", "Species")], function(Genus, Species) {
  rtn = try(fl_mcupd(Genus=Genus, Species=Species)$traits)
  if ("try-error" %in% class(rtn)) return(NULL)
  rtn})

sink(file=NULL)
priors=mdply(sppSiofa[,c("Genus","Species")], function(Genus,Species) {
  
  rtn=try(Plot_taxa(Search_species(Genus=Genus,Species=Species)$match_taxonomy,mfrow=c(3,2)))
  
  if ("try-error"%in%is(rtn)) return(NULL)
  data.frame(trait=names(rtn[[1]][[2]]),data=rtn[[1]][[2]],var=diag(rtn[[1]][[1]]))})
sink()

# Merge priors and traits data
lh=merge(priors, traits, by=c("Genus", "Species", "trait"), all=T)

names(lh)=c("Genus","Species","trait","Value","Variance","Hat","CV","mu.upd","CV.upd","lcl.upd","ucl.upd","upper.quant")
```

```{r}
save(lh,file="../data/cCoelolepis.RData")
```

# Introduction

In this analysis, we update the biological knowledge on *Centroscymnus coelolepis*, using
data from catch data and biological sampling obtained from SIOFA and IEO 

Based on these data we update the life history parameters such as the length-weight 
relationship ($W=aL^b$), length at maturity, and maximum size ($L_{max}$). The analysis 
also evaluates the current status using Length Based Indicators (LBIs).

## Data 

The datasets include length, mass and maturity-at-age; SIOFA data are in `SIOFA_DWS-2023-01-Data.xlsx`, 
and IEO data in `Sharks_biological.Rdata`, `Sharks_logbooks.Rdata`, and `maturity_data.xlsx`.


```{r, data-siofa}
# Load data from Excel file into data frames
cpue=read.xlsx("../data/inputs/SIOFA_DWS-2023-01-Data.xlsx", "dws_data_catch", 
                 startRow=1, detectDates=TRUE, skipEmptyRows=TRUE)
biol=read.xlsx("../data/inputs/SIOFA_DWS-2023-01-Data.xlsx", "dws_data_bio_sampling",
                 startRow=1, detectDates=TRUE, skipEmptyRows=TRUE)

# Adjusting time format
adjust_time_format<-function(time)
  format(as.POSIXct(time*86400, origin="1970-01-01", tz="UTC"), "%H:%M:%S")

cpue$SettingTime=adjust_time_format(cpue$SettingTime)
cpue$HaulingTime=adjust_time_format(cpue$HaulingTime)
cpue$time       =adjust_time_format(cpue$Time)
cpue$year       =year(cpue$Date)
names(cpue)[c(3,6)]=tolower(names(cpue)[c(3,6)])
names(cpue)[c(7:8)]=c("lat","lon")

cpue=subset(cpue,speciesScientificName=="Centroscymnus coelolepis")
biol=subset(biol,speciesScientificName=="Centroscymnus coelolepis")
biol=transform(biol, gender=factor(Sex,labels=c("Female","Male"),levels=c("F","M")))

siofa=list("biol"=biol,"cpue"=cpue,"cSmry"=catchSmry(cpue))
```

```{r, data-ieo}
load("../data/inputs/Sharks_biological.Rdata")
biol=as.data.frame(biol)
names(biol)[c(3,5,6,7,12)]=c("len","mass","sex","mat","species")

load("../data/inputs/Sharks_logbooks.Rdata")
cpue=as.data.frame(cpue)

biol=merge(biol,cpue[,c("cod_set","month","year","Area")],by=c("cod_set"))
names(biol)=tolower(names(biol))

mat=setNames(seq(4),c("A","B","C","D"))
biol$mat[biol$mat%in%names(mat)]=mat[biol$mat[biol$mat%in%names(mat)]]
biol$mat=substr(biol$mat,nchar(biol$mat),1)

biol=transform(biol,spp=factor(sppAbbrv(species)))
biol=transform(biol,mature=ifelse(as.numeric(mat)>=3,1,0))

names(cpue)[]="species"
cpue=subset(cpue,species=="Centroscymnus coelolepis")
biol=subset(biol,species=="Centroscymnus coelolepis")
biol=transform(biol, gender=factor(sex,labels=c("Female","Male"),levels=c("F","M")))
ieo=list("biol"=biol,"cpue"=cpue)
```

```{r, data-save}
save(ieo,siofa,file="../data/pDog.RData")
```


```{r, mat}
pDogMorph=read.xlsx("../data/inputs/maturity_data.xlsx", "morphometric", 
                 startRow=1, detectDates=TRUE, skipEmptyRows=TRUE)
```

```{r}
pDogGonad=read.xlsx("../data/inputs/maturity_data.xlsx", "gonad", 
                 startRow=1, detectDates=TRUE, skipEmptyRows=TRUE)
```

```{r}
mat=rbind.fill(cbind(Type="Morph",pDogMorph),
               cbind(Type="Gonad",pDogGonad))
mat$sex =factor(mat$sex,labels=c("Female","Male"),levels=c("F","M"))
mat$Type=factor(mat$Type)
```

## Life history parameters

By default the life history parameters are downloaded from 'FishBase'. However, these
are sparse as they do not include information by gender.

**Table `r iTab=iTab+1; iTab`** "Life History parameters from `FishBase` for Centroscymnus coelolepis".

```{r, lhist}
#info    =species("Centroscymnus coelolepis")
library(kableExtra)
lhistory=data.frame( 
  Parameter  =c("Maturity (Lm)", "Max Length", "Common Length", "Max Weight"),
  Description=c("101.8 cm (range 95 - 110 cm)", 
                "121 cm TL male/unsexed", 
                "92.0 cm TL male/unsexed", 
                "10.1 kg"),
  Reference  =c("Ref. 122636", "Ref. 122636", "Ref. 247", "Ref. 122636"))  

kable(lhistory)
```


```{r spp, eval=FALSE}
# Load FAO taxa data
load("../data/inputs/fao_taxa.RData")

# Read ASFIS species information
asfis = read.csv("../data/inputs/ASFIS_sp_2021.txt")
names(asfis)[4] = "scientificName"

# Transform fishbase data for compatibility
spp = transform(rfishbase::fishbase, scientificName = paste(Genus, Species))

# Define species of interest for SIOFA
sppSiofa       =c("Centroscymnus coelolepis","Pailona commun","Portugese dogfish","CYO")
sppSiofa       =as.data.frame(t(array(sppSiofa, c(4, length(sppSiofa)/4))))
sppSiofa       =cbind(ldply(strsplit(sppSiofa[,1], " ")), sppSiofa)
names(sppSiofa)=c("Genus", "Species", "scientificName", "french", "name", "FAO")

# Merge species data with fishbase and ASFIS data
sppSiofa = merge(sppSiofa, spp, by=c("Genus", "Species", "scientificName"), all.x=T)
sppSiofa = merge(sppSiofa, asfis[,-(12:13)], by="scientificName", all.x=T)

# Extract life history traits using custom functions
traits = mdply(sppSiofa[,c("Genus", "Species")], function(Genus, Species) {
  rtn = try(fl_mcupd(Genus=Genus, Species=Species)$traits)
  if ("try-error" %in% class(rtn)) return(NULL)
  rtn})

sink(file=NULL)
priors=mdply(sppSiofa[,c("Genus","Species")], function(Genus,Species) {
  
  rtn=try(Plot_taxa(Search_species(Genus=Genus,Species=Species)$match_taxonomy,mfrow=c(3,2)))
  
  if ("try-error"%in%is(rtn)) return(NULL)
  data.frame(trait=names(rtn[[1]][[2]]),data=rtn[[1]][[2]],var=diag(rtn[[1]][[1]]))})
sink()

# Merge priors and traits data
lh=merge(priors, traits, by=c("Genus", "Species", "trait"), all=T)

names(lh)=c("Genus","Species","trait","Value","Variance","Hat","CV","mu.upd","CV.upd","lcl.upd","ucl.upd","upper.quant")

save(lh,file="../data/cCoelolepis.RData")
```


**Table `r iTab=iTab+1; iTab`** Life history parameters and priors from `FishLife`.

```{r lh-tab}
load("../data/cCoelolepis.RData")

kable(lh[,c(3:7)])
```



### Length Weight Relationship

```{r, lw, fig.height=4}
dat   =subset(ieo$biol,!is.na(len)&!is.na(mass))
female=lm(log(mass) ~ log(len), data=subset(dat,mass<50&gender=="Female"))
male  =lm(log(mass) ~ log(len), data=subset(dat,mass<50&gender=="Male"))

eqFemale=paste("W", format(coef(female)[1], digits=3), 
               "L^",format(coef(female)[2], digits=3))
eqMale  =paste("W",  format(coef(male)[1],  digits=3), 
               "L^", format(coef(male)[2],  digits=3))
 
ggplot(subset(dat, mass < 50), aes(x=len, y=mass)) +
  geom_point() +
  geom_smooth(aes(x=len, y=mass, col=gender), method="lm", se=FALSE, formula=y~x) +
  scale_x_log10() +
  scale_y_log10() +
  facet_grid(. ~ gender) +
  xlab("Log Length (cm)") + ylab("Log Mass (Kg)")+
  geom_text(aes(x=50, y=50, label=eqFemale), 
              data=subset(dat, gender=="Female"), hjust=0, vjust=0,
              check_overlap=TRUE, color="red") +
  geom_text(aes(x=50, y=50, label=eqMale), 
              data=subset(dat, gender=="Male"), hjust=0, vjust=0, 
              check_overlap=TRUE, color="blue")+
  scale_color_manual(values=c("red","blue"))
```

**Figure `r iFig=iFig+1; iFig`** Length weight relationship ($W=aL^b$).


### Maturity 

Codes for maturity stages are **Females:** 1: immature; 2: maturing; 3: mature; 4: uterine developing; 5: differentiating; 6: expecting; 7: post-natal, spent; 8: Re-starting development; and **Males:** 1: immature; 2: developing; 3: mature; 4: active


```{r, mat-ieo, fig.height=4}
ggplot(mat)+
  geom_line(aes(len,fitted,col=sex,linetype=Type))+
  xlab("Length (cm)")+ylab("Proportion Mature") 
```

**Figure `r iFig=iFig+1; iFig`** Maturity Ogives, provided by IEO.

#### Fit logistic regression model
```{r, mat-fit, fig.height=4}
model=glm(mature~len+sex+len:sex+len:Type, data=mat, family=binomial)
mat$hat=predict(model,type="response")

ggplot(mat, aes(x=len, y=hat, color=sex, linetype=Type)) + 
  geom_line() + 
  labs(x="Length", y="Predicted Maturity Probability") + 
  scale_y_continuous(limits=c(0, 1))
```


**Figure `r iFig=iFig+1; iFig`** Maturity Ogives, fitted using data provided by IEO.

**Table `r iTab=iTab+1; iTab`** Life history parameters and priors from `FishLife`.

```{r}
names(mat)[8]="Gender"
kable(ddply(mat,.(Gender,Type), with,
      data.frame("l50"=(len[(fitted-0.5)^2==min((fitted-0.5)^2)][1]))))%>%
  kable_styling(full_width = FALSE, position = "left") %>%
  column_spec(1, width = "4em")  %>%
  column_spec(2, width = "6em")  %>%
  column_spec(3, width = "6em") 
```

**Table `r iTab=iTab+1; iTab`** Life history parameters and priors from `FishLife`.

```{r}
l50<-function(model, sex, Type) {
  # Extract coefficients from the model
  coefs     =coef(model)
  intercept =coefs['(Intercept)']
  coefLen   =coefs['len']

    # Adjust coefficients based on sex
  sexEffect =ifelse(sex == "Male", coefs['sexMale'], 0)
  lenSex    =ifelse(sex == "Male", coefs['len:sexMale'], 0)
  
  # Adjust coefficients based on Type
  typeEffect=ifelse(Type == "Morph", coefs['len:TypeMorph'], 0)
  
  # Calculate the adjusted coefficient for len
  coefLen  =coefLen + lenSex + typeEffect
  
  # Calculate the length at maturity=0.5
  l50      =-(intercept + sexEffect) / coefLen
  
  return(l50)}

l50=mdply(expand.grid(Type  =c("Gonad", "Morph"),
                      Gender=c("Female","Male")), function(Gender,Type) 
                        data.frame("L50"=l50(model,Gender,Type)))
kable(l50[,c(2:1,3)],digits=0)%>%
  kable_styling(full_width = FALSE, position = "left") %>%
  column_spec(1, width = "4em")  %>%
  column_spec(2, width = "6em")  %>%
  column_spec(3, width = "6em") 
```

## Catch

```{r, fig.height=4, fig.width=8}
ctc=ddply(siofa[[3]],.(year),with, data.frame(catch=sum(catchWeight,na.rm=T)))
ggplot(ctc)+
  geom_line(aes(year,catch))
```


**Figure `r iFig=iFig+1; iFig`** Time series of aggregated catch biomass.


## Length Frequency Distributions

### IEO
```{r, len-ieo, fig.width=8, fig.height=4}
nobs =ddply(ieo$biol,.(gender,year), with, data.frame(n=length(len)))
ibiol=left_join(ieo$biol, nobs)

ggplot(ibiol, aes(len)) +
  geom_histogram(bins=50, fill="skyblue", color="blue") +
  coord_flip()+
  facet_grid(gender~year, scales="free_y") +
  #scale_y_continuous(breaks=function(x) {c(0,round(max(x,na.rm=TRUE)))})+  
  geom_text(data=nobs, aes(label=paste("n =", n), x=Inf, y=Inf), 
            position=position_nudge(x=-0.5,y=-0.5), hjust=1, vjust=1)+
  xlab("Length (cm)")+ylab("")+ 
  theme(legend.position="bottom", 
                    axis.title.x=element_blank(), 
                    axis.text.x =element_blank(), 
                    axis.ticks.x=element_blank())
```

**Figure `r iFig=iFig+1; iFig`** Length distributions by gender for IEO data.

```{r, flen}
iflen=ddply(transform(ieo$biol,len=as.integer(len)),.(gender,year,len), 
           with,data.frame(data=length(cod_set)))
iflen=as.FLQuant(mutate(iflen,unit=gender)[,-1])
iflen[is.na(iflen)]=0
```

### SIOFA


```{r,len-siofa, fig.width=8, fig.height=4}
sbiol=merge(siofa$biol[,c(1,11:12,16)],siofa$cpue[,c("OPE_ID","year")])[,-1]
names(sbiol)=c("len","mass","unit","year")
sbiol=subset(sbiol,!is.na(len)&unit%in%c("Male","Female"))
nobs =ddply(sbiol,.(unit,year), with, data.frame(n=length(len)))

sflen=ddply(sbiol,.(len,year,unit), 
           with,data.frame(data=length(mass)))
sflen=as.FLQuant(sflen)
sflen[is.na(sflen)]=0

ggplot(sbiol, aes(len)) +
  geom_histogram(bins=50, fill="skyblue", color="blue") +
  coord_flip()+
  facet_grid(unit~year, scales="free_y") +
  #scale_y_continuous(breaks=function(x) {c(0,round(max(x,na.rm=TRUE)))})+
  geom_text(data=nobs, aes(label=paste("n =", n), x=Inf, y=Inf), 
            position=position_nudge(x=-0.5,y=-0.5), hjust=1, vjust=1)+
  xlab("Length (cm)")+ylab("")+ 
  theme(legend.position="bottom", 
                    axis.title.x=element_blank(), 
                    axis.text.x =element_blank(), 
                    axis.ticks.x=element_blank())+
  geom_vline(aes(xintercept=data),data=as.data.frame(lc50( sflen)),col="red")+
  geom_vline(aes(xintercept=data),data=as.data.frame(lmean(sflen)),col="black")
```

**Figure `r iFig=iFig+1; iFig`** Length distributions by gender for SIOFA data.


### Maximum size

The empirical relationship between $L_{\infty}$ (asymptotic length, the maximum length towards which the individuals in a population converge as they become infinitely old) and $L_{max}$ (the maximum observed length in a population) is important in data limited stocks for assessing fish growth and developing priors for population dynamics.

While both $L_{\infty}$ and $L_{max}$ provide insights into the growth potential of a species, they are distinct parameters. $L_{\infty}$ is a theoretical parameter derived from growth models (e.g., the von Bertalanffy growth function), whereas $L_{max}$ is an empirical observation from field data.

Several studies have attempted to establish empirical relationships between these two metrics to facilitate the estimation of one parameter from the other, especially in data-poor situations. However, it's important to note that any empirical relationship may be species-specific and influenced by environmental factors, fishing pressure, and the methodological approaches used in studies. A commonly cited rule of thumb is that $L_{max}$ is often observed to be approximately 0.95 to 0.99 times $L_{\infty}$, acknowledging that most individuals in a population do not reach or exceed the asymptotic length due to natural mortality, predation, or fishing pressure. This approximation allows researchers to estimate $L_{\infty}$ from $L_{max}$ when growth model parameters are unknown or cannot be directly calculated due to lack of data. 

However, relying on a fixed ratio or empirical relationship without considering species-specific life history traits, ecological conditions, and the quality of the available data can lead to inaccuracies. Therefore, when such empirical relationships are used, they should be applied with caution and, if possible, validated against more comprehensive data sets or models specific to the species and region of interest.

**Table `r iTab=iTab+1; iTab`** .
```{r,lmax-i}
kable(lmax(iflen)[drop=T],col.names = c("Female","Male")) %>% 
  kable_styling(full_width = FALSE, position = "left") %>%
  column_spec(1, width = "4em")  %>%
  column_spec(2, width = "6em")  %>%
  column_spec(3, width = "6em") 
```

**Table `r iTab=iTab+1; iTab`** .
```{r,lmax-s}
kable(lmax(sflen)[drop=T],col.names = c("Female","Male")) %>% 
  kable_styling(full_width = FALSE, position = "left") %>%
  column_spec(1, width = "4em")  %>%
  column_spec(2, width = "6em")  %>%
  column_spec(3, width = "6em") 
```
